separated the reusable collapsible card and progressbar components
