+++ gauchograduate/src/app/components/progress_components/CollapsibleCard.tsx
@@ -1,0 +1,40 @@
+// CollapsibleCard.tsx
+import React, { useState } from "react";
+
+export interface CollapsibleCardProps {
+  title: string;
+  children: React.ReactNode;
+}
+
+const CollapsibleCard: React.FC<CollapsibleCardProps> = ({ title, children }) => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  return (
+    <div className="bg-white rounded-lg shadow p-4 mb-4">
+      <div
+        className="flex justify-between items-center cursor-pointer"
+        onClick={() => setIsOpen(!isOpen)}
+      >
+        <h3 className="text-lg font-semibold">{title}</h3>
+        <svg
+          className={`w-6 h-6 transform transition-transform ${
+            isOpen ? "rotate-180" : "rotate-0"
+          }`}
+          fill="none"
+          stroke="currentColor"
+          viewBox="0 0 24 24"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M19 9l-7 7-7-7"
+          />
+        </svg>
+      </div>
+      {isOpen && <div className="mt-4">{children}</div>}
+    </div>
+  );
+};
+
+export default CollapsibleCard;
